refactor(platform): extract shared CLI machinery into config/cli.rs (#2830)
Move CliFlag, ParseResult, the central PLATFORM_CLI_FLAGS table,
OPTIONAL_BOOL_FLAGS, generic value parsers, all_cli_flags, validate_args,
and help text generation into a new config/cli.rs. mod.rs keeps the data
structs and re-exports the externally-used items so crate::platform::config::*
paths are unchanged. No behavior change.
